Added double slider reads threshold to the DDA Widget. Refactor DDA Widget. Fix a bug that would case the pooling mode and color mode to not reset.

+++ src/analysis/AnalysisDEA.h
@@ -74,7 +74,7 @@ class AnalysisDEA : public QDialog
 
 public:
 
-    typedef QMap<QString, QPair<qreal,qreal> > geneToReadsPairType;
+    typedef QMap<QString, QPair<int,int> > geneToReadsPairType;
 
     AnalysisDEA(const GeneSelection& selObjectA,
                 const GeneSelection& selObjectB,
